Move createGIFTexture to gif-texture.js
